feat: switch to light theme and clarify env errors
