+++ netlify/functions/supabase.ts
@@ -6,23 +6,26 @@ import type { Handler, HandlerEvent, HandlerContext } from "@netlify/functions";
  * It fetches the top 1 record from a 'notes' table.
  */
 const handler: Handler = async (event: HandlerEvent, context: HandlerContext) => {
-  const { SUPABASE_URL, SUPABASE_ANON_KEY } = process.env;
+  const { SUPABASE_URL, SUPABASE_SERVICE_ROLE_KEY } = process.env;
 
   const headers = {
     'Content-Type': 'application/json',
     'Access-Control-Allow-Origin': '*', // Allow requests from any origin
   };
 
-  if (!SUPABASE_URL || !SUPABASE_ANON_KEY) {
+  const missing: string[] = [];
+  if (!SUPABASE_URL) missing.push('SUPABASE_URL');
+  if (!SUPABASE_SERVICE_ROLE_KEY) missing.push('SUPABASE_SERVICE_ROLE_KEY');
+  if (missing.length) {
     return {
       statusCode: 500,
-      body: JSON.stringify({ error: 'Supabase environment variables (SUPABASE_URL, SUPABASE_ANON_KEY) are not set in the Netlify dashboard.' }),
+      body: JSON.stringify({ error: `Missing environment variables: ${missing.join(', ')}` }),
       headers,
     };
   }
 
   try {
-    const supabase = createClient(SUPABASE_URL, SUPABASE_ANON_KEY);
+    const supabase = createClient(SUPABASE_URL, SUPABASE_SERVICE_ROLE_KEY);
 
     // Perform a simple query to test the connection and credentials
     const { data, error } = await supabase
